+++ test/runners/env/setup_debug_env.sh
@@ -7,6 +7,7 @@ printf "\n*\n* Running in DEBUG MODE, sourcing setup files\n*\n"
 if . test/runners/env/set_reuse_tmp_env.sh; then
    . test/runners/env/test_env.sh     # set dataset, pools and remote env vars
    . test/test_helper.sh          # make all the helper functions available
+   env | grep -i sandbox
 else
     return 1
 fi