+++ test/runners/test_generation/bin/debug_gen.sh
@@ -1,44 +1,29 @@
-# This is a helper for debugging and testing your generated tests
-# This can be helpful if you use the generate_new_tests.sh approach
-# and your test confirmation fails.
-#
-# Why? Because you can fine-tune the tree setup and run your test commands
-# by hand to determine the cause of problems. Your test yml definition may
-# need additional setup or a modified command. So the ability to iteratively
-# test out the new spec can help you resolve problems.
-#
+#!/usr/bin/env bash
+# Get the directory where this script is located
+SCRIPT_DIR=$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)
+
 # SPECS    - the shellspec examples you run to setup the tree before running your test
-# NEW_SPEC - the generated test you are debugging or verifying
-
-REPO_ROOT=$(git rev-parse --show-toplevel)
-
-# standard locations under test
-RUNNERS_DIR="$REPO_ROOT/test/runners"
-TEST_GEN_DIR="$REPO_ROOT/test/runners/test_generation"
-
-# tree setup utility
-. "$TEST_GEN_DIR/lib/orchestration/setup_tree.sh"
-
 # prepare the zfs tree with the state represented by running the following examples/tests
-SPECS="test/01*_spec.sh|test/01*_spec.sh|test/02*_spec.sh|test/040_*_spec.sh"
+# TODO: change the SPECS variable to include all tests you need to run before the new test
+SPECS="test/01*_spec.sh|test/02*_spec.sh|test/03_*_spec.sh|test/040_*_spec.sh|test/050_*_spec.sh|test/060_*_spec.sh"
 
-# use the directory where your generated test is created
-# by default we're using a temp directory off of test/runners/test_generation
-NEW_SPEC="$TEST_GEN_DIR/tmp/050_zelta_revert_spec.sh"
-echo "confirming new spec: {$NEW_SPEC}"
-
-if setup_tree "$SPECS"; then
-    printf "\n ✅ initial tree setup succeeded for specs: %s\n" "$SPECS"
-else
-    printf "\n ❌ Goodbye, initial tree setup failed for specs: %s\n" "$SPECS"
+if ! . "$SCRIPT_DIR/setup_debug_state.bash" "$SPECS"; then
+    printf "\n ❌ ERROR: debug state setup failed\n"
     exit 1
 fi
 
-#
-. "$RUNNERS_DIR/env/setup_debug_env.sh"
+# use the directory where your generated test is created
+# by default we're using a temp directory off of test/runners/test_generation
+# TODO: change the NEWS_SPEC variable to be spec of the new test you debugging
+NEW_SPEC="$TEST_GEN_DIR/tmp/070_zelta_prune_spec.sh"
+echo "confirming new spec: {$NEW_SPEC}"
+
 
 # show a detailed trace of the commands you are executing in your new test
-TRACE_OPTIONS="--xtrace --shell /opt/homebrew/bin/bash"
+# macOS BASH_SH=/opt/homebrew/bin/bash
+# Arch  BASH_SH=/usr/bin/bash
+BASH_SH=/usr/bin/bash
+TRACE_OPTIONS="--xtrace --shell $BASH_SH"
 # if you don't want/need a detailed trace unset the options var
 #unset TRACE_OPTIONS