+++ test/runners/test_generation/lib/orchestration/generate_test.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/usr/bin/env bash
 
 # Check for required arguments
 if [ $# -lt 2 ]; then